+++ android/build.gradle
@@ -105,14 +105,14 @@ dependencies {
 
   implementation 'com.facebook.react:react-native:'
   
-  implementation 'com.yandex.div:div-core:29.11.0'
-  implementation 'com.yandex.div:div:29.11.0'
-  implementation 'com.yandex.div:div-json:29.11.0'
+  implementation 'com.yandex.div:div-core:30.3.0'
+  implementation 'com.yandex.div:div:30.3.0'
+  implementation 'com.yandex.div:div-json:30.3.0'
 
-  implementation 'androidx.core:core-ktx:1.12.0'
-  implementation 'androidx.appcompat:appcompat:1.6.1'
+  implementation 'androidx.core:core-ktx:1.13.1'
+  implementation 'androidx.appcompat:appcompat:1.7.0'
   implementation 'androidx.constraintlayout:constraintlayout:2.1.4'
   implementation "com.squareup.picasso:picasso:2.8"
-  implementation "com.squareup.okhttp3:okhttp:4.9.2"
+  implementation "com.squareup.okhttp3:okhttp:4.10.0"
 
 }
